+++ docs/changelog.rst
@@ -14,6 +14,7 @@ Changelog
 * Runtime: Android: added support for keyboard, mouse and joypad input devices.
 * Runtime: HTML5: added support for joypad input device.
 * Runtime: added the ability to select which GPU to boot the renderer with using the new ``renderer.device_id`` field of :doc:`boot.config <reference/boot_config>`.
+* Lua: added ``PhysicsWorld.mover_actor_colliding_down()``.
 * Lua: added ``mover_actor_collision()`` and ``mover_mover_collision()`` script component callbacks.
 
 **Fixes**